+++ rest-account/src/main/resources/static/.yarnrc.yml
