+++ rest-account/src/main/resources/static/README.md
