+++ rest-account/src/main/resources/static/asset/favicon.ico
