+++ rest-account/src/main/resources/static/load-webpack-files.js
