+++ rest-account/src/main/resources/static/package.json
