+++ rest-account/src/main/resources/static/webpack.config.js
